feat: support Windows hosts (#47)
Closes #44

Completes cross-platform support (Linux + macOS landed in #46): podup
now builds, tests and releases for Windows.

## What

- **Named pipe connection** (`internal/podman/mod.rs`): on Windows podup
connects through `Docker::connect_with_named_pipe`, resolving the pipe
via `podman machine inspect --format '{{ .ConnectionInfo.PodmanPipe.Path
}}'` and falling back to the default machine's
`//./pipe/podman-machine-default`. `DOCKER_HOST`/`PODMAN_SOCKET` values
with an `npipe://` prefix are honored.
- **Staging extracted to `engine::staging`** (new module, also shrinks
the 799-line `volume.rs` per the file-size standard): all
platform-specific filesystem behavior for inline secret/config
materialisation now lives in one place.
- unix: byte-for-byte the previous behavior — verified 0700
`XDG_RUNTIME_DIR` base, atomic 0600 file creation, mode/chown
application, fail-closed ownership checks. Tests moved along.
- Windows: stages under `%TEMP%\podup` (per-user ACLs by default, unlike
world-writable `/tmp`; symlink/non-dir still rejected) and warns when
compose `mode:`/`uid:`/`gid:` fields cannot apply.
- **libc** becomes a `[target.'cfg(unix)'.dependencies]` entry.
- **CI**: `windows-latest` joins `extra-test-os`.
- **Release**: `podup-windows-x86_64.exe` (+`.sig`, attestation,
checksum) joins the matrix; build steps standardized on bash across
runners. `install.sh` stays unix-only — the README points Windows users
at the release asset.

## Test plan

- `cargo test --workspace` — 264 tests green on Linux
- `cargo clippy --all-targets --all-features` clean on
`x86_64-unknown-linux-gnu`, `x86_64-pc-windows-gnu` and
`aarch64-apple-darwin`
- `cargo check --target x86_64-pc-windows-gnu --all-targets` — whole
crate, tests included
- The `Test (windows-latest)` and `Test (macos-latest)` CI jobs on this
PR run the suite on real runners
- Real `podman machine` smoke test on Windows hardware pending — tracked
with the macOS smoke test before tagging the next release

---------
